refactor: split matrix.client_visible_messages and matrix.client_thread_history (ARCH-B-3 step 1)
